Add files via upload
I'm new to GitHub so I don't really know how to do this. I seem to have managed to fix all of the syntax errors with the small ruins. In my tests I am not getting any crashes with all of them enabled. I may make future tweaks to them but this is the rough draft.